requirements.txt
# Force redeploy
Flask==3.0.0
Werkzeug==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
cloudinary==1.36.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
flask-mail==0.10.0
wtforms==3.1.0
email-validator==2.1.0
authlib==1.3.1
requests==2.31.0
python-docx==1.1.0
boto3==1.34.0
supabase==2.4.6
